Getting the Github author login by searching a commit.
The current implementation searches the user based on his `git` email,
which means that if a user as a different email configured in his local
`.git/config` file and in his primary email on Github we will fail to
find him based on his email.

This commit is about to solve this issue but searching for the commit
that was introduced in a PR and extracting the commit author from there.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>