Adding the `GOCOVERDIR` to the mock of one of the tests.
When running tests with the `-coverprofile` flag, the go binary will try to
collect coverage for the test.

Since this test is mocking a command, then the `-coverprofile cover.out`
of the `go test` command will be ignored, therefore, it order to make it
work we need to set the `GOCOVERDIR` directory in the mock.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>